backend/requirements.txt
Django>=5.2.1
mongoengine>=0.27.0
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
PyJWT>=2.8.0
dnspython>=2.4.2